+++ src/components/NotificationSettings/NotificationConfigure/Channel.scss
@@ -1,5 +1,6 @@
 .channel-container {
-    
+    position: relative;
+    right: 20px; 
     .logocompany {
         min-width: 40px;
         max-height: 40px;
@@ -11,13 +12,13 @@
         top: 2px;
         border-radius: 5px !important;
       }
-  
-  
+
+
       .bx-shadow {
         padding: 10px;
         background: #fff !important;
-        margin-bottom: 10px; 
-        
+        margin-bottom: 10px;
+
         &:hover {
           box-shadow: 0px 12px 16px -5px rgba(0, 0, 0, 0.2) !important;
           background-color: rgb(255, 255, 255) !important;
@@ -34,10 +35,10 @@
           margin-bottom: 50px !important;
 
       }
-  
-   
-  
-  
-  
-  
+
+
+
+
+
+
 }
