+++ src/components/NotificationSettings/NotificationConfigure/NotificationConfigure.js
@@ -4,6 +4,7 @@ import { Card } from './Card';
 import NotificationDesign from './NotificationDesign';
 import { NotificationDesignSetting } from './NotificationDesignSetting';
 import './NotificationConfigure.scss';
+// import PopupReview from './PopupReview';
 
 const NotificationConfigure = ({
   profile,
@@ -21,10 +22,13 @@ const NotificationConfigure = ({
   handleNotificationStyleChange,
   handleClickableNotification,
   saveConfiguration,
-  backConfiguration
+  backConfiguration,
+  showpopupfield,
+  showpopup,
+  popupName
 }) => {
   return (
-    <div className="notification-configure">
+    <div className="notification-configure col-md-12">
       <Row>
         <Col md={12}>
           <Card title={notification.notificationName}
@@ -50,11 +54,15 @@ const NotificationConfigure = ({
                       visitorText={visitorText}
                       notification={notification}
                       notificationPanelStyle={notificationPanelStyle}
+                      popupName={popupName}
                     />
                   </Col>
                   <Col md={5}>
+
                     <NotificationDesignSetting
                       profile={profile}
+                      showpopup={showpopup}
+                      showpopupfield={showpopupfield}
                       toggleTextBox={toggleTextBox}
                       toggleMap={toggleMap}
                       contentText={contentText}
@@ -66,6 +74,7 @@ const NotificationConfigure = ({
                       handleContentChange={handleContentChange}
                       handleClickableNotification={handleClickableNotification}
                     />
+
                   </Col>
                 </Row>
               </div>
@@ -74,22 +83,21 @@ const NotificationConfigure = ({
         </Col>
       </Row>
       <Row className="state-btn">
-        <Col md={4}>
 
-          <span className="btn btn-primary mr-3" onClick={backConfiguration}>
+        <span className="btn btn-primary" onClick={backConfiguration}>
           Back
-          </span>
-        </Col>
-        <Col md={4}>
-          <span className="btn btn-primary mr-3" onClick={setDefaultPanel} >
+        </span>
+
+
+        <span className="btn btn-primary" onClick={setDefaultPanel} >
           Set Default
-          </span>
-        </Col>
-        <Col md={4}>
-          <span className="btn btn-primary mr-3 " onClick={saveConfiguration}>
+        </span>
+
+
+        <span className="btn btn-primary" onClick={saveConfiguration}>
           Save
-          </span>
-        </Col>
+        </span>
+
 
       </Row>
     </div>
