+++ src/components/NotificationSettings/NotificationConfigure/NotificationDesign.js
@@ -1,4 +1,10 @@
 import React from 'react';
+import {
+  Googlepop,
+  Facebookpop,
+} from 'img';
+import StarRatings from 'react-star-ratings';
+import './NotificationDesign.scss';
 
 const NotificationDesign = ({
   contentText,
@@ -6,6 +12,7 @@ const NotificationDesign = ({
   toggleMap,
   notificationPanelStyle,
   notification,
+  popupName
 }) => {
   const {
     radius,
@@ -38,6 +45,8 @@ const NotificationDesign = ({
   };
 
   const notificationName = notification.notificationName;
+  let img={};
+  popupName === 'Google' ? img = Googlepop : img= Facebookpop;
 
   return (
     <div className="card-box p-0">
@@ -203,10 +212,80 @@ const NotificationDesign = ({
                           </div>
                         </div>
                       </div>
+                      <div style={notificationName == 'Review Notification'
+                        ? {}
+                        : {
+                          display: 'none'
+                        }}>
+                        <div className="reviewpop-container">
+                          <div id="FPqR2fZIqJeA2fZI7MM9_0">
+                            <div className="FPqR3zjZqJeA3zjZ7MM9_0 FPqR2riIqJeA2riI7MM9_0 wrapper"  style={{
+                              borderRadius: notificationStyle.borderRadius,
+                              backgroundColor: notificationStyle.backgroundColor,
+                              border: `${notificationStyle.borderWidth} ${notificationStyle.borderStyle} ${notificationStyle.borderColor}`
+                            }}>
+
+                              <div className="FPqR1JYFqJeA1JYF7MM9_0" ><img src={img} width='50px'/></div>
+                              <div className="FPqR2EbCqJeA2EbC7MM9_0 content-pop">
+                                <div className="FPqR2AUlqJeA2AUl7MM9_0" style ={{
+                                  color:notificationStyle.color,
+                                  fontFamily: notificationStyle.fontFamily,
+                                  fontWeight: notificationStyle.fontWeight}}>
+                                  <div className="star">
+                                    <StarRatings
+                                      rating={4.6}
+                                      starDimension="20px"
+                                      starRatedColor={popupName === 'Google' ? 'gold' : '#097fff'}
+                                      numberOfStars={5}
+                                      starSpacing= "5px"
+                                      name='rating'
+                                    />
+                                    <span  className="rating-num" style ={{
+                                      fontFamily: notificationStyle.fontFamily,
+                                      fontWeight: notificationStyle.fontWeight}}> 4.6</span></div>
+                                </div>
+                                <div className="FPqR2PlWqJeA2PlW7MM9_0 "  style={{
+                                  color: notificationStyle.linkColor,
+                                  fontFamily: notificationStyle.linkFontFamily,
+                                  fontWeight: notificationStyle.linkFontWeight,
+                                  backgroundColor:notificationStyle.backgroundColor
+                                }}>
+                                  <span className="FPqR1Jr6qJeA1Jr67MM9_0 headtext-pop">
+                                    <span>250 {visitorText} reviewed {contentText}</span>
+                                  </span>
+
+                                </div>
+                                <div className="FPqR3eNuqJeA3eNu7MM9_0 verifier">
+                                  <i>
+                                    <svg width="7" height="13" viewBox="0 0 7 13" xmlns="http://www.w3.org/2000/svg">
+                                      <g fill="none" fillRule="evenodd">
+                                        <path d="M4.127.496C4.51-.12 5.37.356 5.16 1.07L3.89 5.14H6.22c.483 0 .757.616.464 1.044l-4.338 6.34c-.407.595-1.244.082-1.01-.618L2.72 7.656H.778c-.47 0-.748-.59-.48-1.02L4.13.495z" fill="#F6A623"></path>
+                                        <path fill="#FEF79E" d="M4.606.867L.778 7.007h2.807l-1.7 5.126 4.337-6.34H3.16"></path>
+                                      </g>
+                                    </svg>
+                                  </i>
+                                  <span>by  </span>
+                                  <span className="FPqR12wMqJeA12wM7MM9_0" style={{
+                                    color: notificationStyle.linkColor,
+                                    fontFamily: notificationStyle.linkFontFamily,
+                                    fontWeight: notificationStyle.linkFontWeight
+                                  }}><b>Influence</b></span>
+                                </div>
+                              </div>
+
+                              <div className="FPqR1N_EqJeA1N_E7MM9_0" style={{display: 'none'}}>
+                                <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
+                                  <path d="M12 0c-6.627 0-12 5.373-12 12s5.373 12 12 12 12-5.373 12-12-5.373-12-12-12zm4.597 17.954l-4.591-4.55-4.555 4.596-1.405-1.405 4.547-4.592-4.593-4.552 1.405-1.405 4.588 4.543 4.545-4.589 1.416 1.403-4.546 4.587 4.592 4.548-1.403 1.416z">
+                                  </path>
+                                </svg>
+                              </div>
+                            </div>
+                          </div>
+                        </div>
+                      </div>
                     </div>
                   </div>
                 </div>
-                <div></div>
               </div>
             </div>
           </div>
