+++ src/components/NotificationSettings/NotificationConfigure/NotificationDesign.scss
@@ -22,12 +22,12 @@
           position: relative;
       }
       .rating-num {
-          position: absolute;   
+          position: absolute;
           right:20px;
           top: 2px;
           font-size: 21px;
           font-weight: 800;
-          
+
 
         }
         .content-pop {
@@ -43,7 +43,7 @@
             top: 5px;
         }
         .FPqR3eNuqJeA3eNu7MM9_0 {
-           
+
             &.verifier {
              position: relative;
              top: 5px;
@@ -53,7 +53,7 @@
         .star {
             position: relative;
             top: -1px;
-         
+
         }
         .review-text {
             font-size: 10px !important;
