+++ src/components/NotificationSettings/NotificationConfigure/NotificationDesignSetting.scss
@@ -1,0 +1,10 @@
+.DashboardChannelcomponent {
+  position: absolute;
+  top: -80px;
+  left: 240px;
+}
+.box1 ,.box2{
+  width: 500px !important;
+  position: relative;
+  right: 10px;
+}
