+++ src/components/NotificationSettings/NotificationConfigure/PopupReview.scss
@@ -1,7 +1,7 @@
 .review-container {
   padding-bottom: 100px;
   height: 680px !important;
- 
+
   .review-title{
     position: relative;
   }
@@ -18,7 +18,7 @@
       font-family: "Rubik", sans-serif;
     }
 
-   
+
   }
   a {
     &.active.pb-3.pt-3 {
@@ -33,7 +33,6 @@
     position: relative;
     top: 200px;
     left: 50px;
-   
+
   }
 }
-
