+++ src/components/NotificationSettings/NotificationList/NotificationList.js
@@ -1,12 +1,12 @@
 import React from 'react';
-import { browserHistory } from 'react-router';
 import './NotificationList.scss';
 
 
 const NotificationList = ({ notificationList, configure, handleActivityChange}) => {
   const renderNotifications = () => {
     return notificationList.map(notification =>
-      <div key={notification._id} className="col-md-4 notification-box">
+      <div key={notification._id}
+        className={notification.notificationName === 'Review Notification' ? 'col-md-4 notification-box lastbox' : 'col-md-4 notification-box'}>
         <div className=" ribbon-box bx-shadow">
           <div className="ribbon-two ribbon-two-success">
             <span>
@@ -17,7 +17,10 @@ const NotificationList = ({ notificationList, configure, handleActivityChange}) 
                   notification.notificationName === 'Live Visitor Count' ?
                     'LIVE'
                     :
-                    'RECENT'
+                    notification.notificationName === 'Recent Activity' ?
+                      'RECENT'
+                      :
+                      'REVIEWS'
               }
             </span>
           </div>
@@ -31,7 +34,10 @@ const NotificationList = ({ notificationList, configure, handleActivityChange}) 
                     notification.notificationName === 'Live Visitor Count' ?
                       'liveColor'
                       :
-                      'recentColor'
+                      notification.notificationName === 'Recent Activity' ?
+                        'recentColor'
+                        :
+                        'reviewColor'
                 }>{notification.notificationName}</h4>
                 <p className="text-muted notification-list-content">
                   {
@@ -41,7 +47,11 @@ const NotificationList = ({ notificationList, configure, handleActivityChange}) 
                       notification.notificationName === 'Live Visitor Count' ?
                         'Live visitors.'
                         :
-                        'Recent visitors.'
+                        notification.notificationName === 'Recent Activity' ?
+                          'Recent visitors.'
+                          :
+                          'Get popup Insights'
+
                   }
                 </p>
               </div>
@@ -55,14 +65,23 @@ const NotificationList = ({ notificationList, configure, handleActivityChange}) 
                 />
                 <label className="tgl-btn" htmlFor={notification._id}></label>
               </div>
-
-              <button
-                type="button"
-                className="btn btn-primary m-t-20 btn-rounded btn-bordered waves-effect w-md waves-light"
-                onClick={() => configure(notification)}
-              >
+              {notification.notificationName == 'Review Notification' ?
+                <button
+                  type="button"
+                  className="btn btn-primary m-t-20 btn-rounded btn-bordered waves-effect w-md waves-light"
+                  onClick={() => configure(notification,false)}
+                >
                 Configure
-              </button>
+                </button>
+                :
+                <button
+                  type="button"
+                  className="btn btn-primary m-t-20 btn-rounded btn-bordered waves-effect w-md waves-light"
+                  onClick={() => configure(notification,true)}
+                >
+                Configure
+                </button>
+              }
             </div>
           </div>
         </div>
@@ -74,41 +93,6 @@ const NotificationList = ({ notificationList, configure, handleActivityChange}) 
     <div className="cards-notification">
       <div className="row m-t-50 notification-list">
         {renderNotifications()}
-      </div>
-      <div className="col-md-4 notification-box addChannel">
-        <div className=" ribbon-box bx-shadow">
-          <div className="ribbon-two ribbon-two-success">
-            <span>
-               Ratings
-            </span>
-          </div>
-        </div>
-        <div className="text-center card-box bx-shadow ">
-          <div className="member-card">
-            <div>
-              <h4 className='channelColor'>Review Popups</h4>
-              <p className="text-muted notification-list-content">Get Popups Insights.</p>
-            </div>
-            <div className="toggle-checkbox">
-              <input
-                className="tgl tgl-ios"
-                id={'channel+'}
-                type="checkbox"
-                checked={true} // need to chenge this too
-                readOnly //need to add on change event
-              />
-              <label className="tgl-btn" htmlFor={'channel+'}></label>
-            </div>
-
-            <button
-              type="button"
-              className="btn btn-primary m-t-20 btn-rounded btn-bordered waves-effect w-md waves-light"
-              onClick={()=>browserHistory.push('/popupreview')}
-            >
-                Configure
-            </button>
-          </div>
-        </div>
       </div>
     </div>
   );
