+++ src/components/index.js
@@ -47,9 +47,6 @@ export { default as DisplayPage } from './DisplayPage/DisplayPage';
 export { default as CapturePage } from './CapturePage/CapturePage';
 export { default as InstallPixel } from './InstallPixel/InstallPixel';
 
-export { default as PopupReview } from './PopupReview/PopupReview';
 export { default as Oauthpage } from './Oauthpage/Oauthpage';
 export { default as Oauthgenerate } from './Oauthgenerate/Oauthgenerate';
 export { default as DashboardChannel  }  from './DashboardChannel/DashboardChannel';
-
-
